[FLINK-22790] Move all classes to the same package
