+++ flink-end-to-end-tests/flink-end-to-end-tests-common-kafka/src/test/java/org/apache/flink/tests/util/kafka/hybrid/FileDataWriter.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.flink.connector.kafka.source.testutils.hybrid;
+package org.apache.flink.tests.util.kafka.hybrid;
 
 import org.apache.flink.connectors.test.common.external.SourceSplitDataWriter;
